+++ CHANGELOG.md
@@ -9,6 +9,8 @@ CHANGELOG
   classes rather than magic methods. This significantly improves performance.
 * BREAKING: The `raw` property on model classess and the `record` property on
   record classes have been removed.
+* BREAKING: On `GeoIp2\Record\Traits`, the deprecated `isAnonymousProxy` and
+  `isSatelliteProvider` properties have been removed.
 * BREAKING: The `jsonSerialize` output has changed.
 * `GeoIp2\WebService\Client` methods now throw an `InvalidArgumentException`
   if an invalid IP address is passed to them. Previously, they would make